+++ runzi/runners/run_oq_disagg.py
@@ -11,7 +11,7 @@ from runzi.automation.scaling.schedule_tasks import schedule_tasks
 from runzi.automation.scaling.toshi_api import ModelType, SubtaskType
 from runzi.configuration.openquake.oq_disagg import build_disagg_tasks
 
-from .config import DisaggInput
+from .hazard_inputs import DisaggInput
 
 loglevel = logging.INFO
 logging.basicConfig(level=logging.INFO)
