+++ runzi/runners/run_oq_hazard.py
@@ -14,7 +14,7 @@ from runzi.automation.scaling.schedule_tasks import schedule_tasks
 from runzi.automation.scaling.toshi_api import CreateGeneralTaskArgs, ModelType, SubtaskType, ToshiApi
 from runzi.configuration.openquake.oq_hazard import build_hazard_tasks
 
-from .config import HazardInput
+from .hazard_inputs import HazardInput
 
 loglevel = logging.INFO
 logging.basicConfig(level=logging.INFO)
